add userFactory
