+++ pages/api/logout.ts
@@ -1,11 +1,11 @@
 import { withIronSessionApiRoute } from "iron-session/next";
 import { sessionOptions } from "../../lib/session";
 import { NextApiRequest, NextApiResponse } from "next";
-import type { User } from "./user";
+import userFactory, { User } from "../../lib/userFactory";
 
 function logoutRoute(req: NextApiRequest, res: NextApiResponse<User>) {
   req.session.destroy();
-  res.json({ uid: -1, username: "", password: "", isLoggedIn: false });
+  res.json(userFactory());
 }
 
 export default withIronSessionApiRoute(logoutRoute, sessionOptions);
