+++ pages/api/user.ts
@@ -1,17 +1,11 @@
 import { withIronSessionApiRoute } from "iron-session/next";
 import { sessionOptions } from "../../lib/session";
 import { NextApiRequest, NextApiResponse } from "next";
-
-export type User = {
-  uid: number;
-  username: string;
-  password: string;
-  isLoggedIn: boolean;
-};
+import userFactory, { User } from "../../lib/userFactory";
 
 async function userRoute(req: NextApiRequest, res: NextApiResponse<User>) {
   if (!req.session.user) {
-    return res.json({ uid: -1, username: "", password: "", isLoggedIn: false });
+    return res.json(userFactory());
   }
   res.json({
     ...req.session.user,
